+++ sdk/spring/spring-cloud-azure-resourcemanager/src/main/java/com/azure/spring/resourcemanager/provisioning/eventhubs/DefaultEventHubsProvisioner.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.spring.resourcemanager.provisioner.eventhubs;
+package com.azure.spring.resourcemanager.provisioning.eventhubs;
 
 import com.azure.resourcemanager.AzureResourceManager;
 import com.azure.spring.core.properties.resource.AzureResourceMetadata;