+++ sdk/spring/spring-cloud-azure-service/src/test/java/com/azure/spring/service/implementation/storage/queue/TestAzureStorageQueueProperties.java
@@ -32,4 +32,5 @@ class TestAzureStorageQueueProperties extends TestAzureStorageProperties impleme
     public void setMessageEncoding(QueueMessageEncoding messageEncoding) {
         this.messageEncoding = messageEncoding;
     }
+
 }