+++ sdk/spring/spring-messaging-azure-storage-queue/src/main/java/com/azure/spring/storage/queue/core/StorageQueueTemplate.java
@@ -6,8 +6,9 @@ package com.azure.spring.storage.queue.core;
 import com.azure.spring.messaging.AzureHeaders;
 import com.azure.spring.messaging.PartitionSupplier;
 import com.azure.spring.messaging.checkpoint.AzureCheckpointer;
-import com.azure.spring.messaging.checkpoint.CheckpointMode;
 import com.azure.spring.messaging.checkpoint.Checkpointer;
+import com.azure.spring.messaging.core.SendOperation;
+import com.azure.spring.storage.queue.core.factory.StorageQueueClientFactory;
 import com.azure.spring.storage.queue.support.StorageQueueHelper;
 import com.azure.spring.storage.queue.support.converter.StorageQueueMessageConverter;
 import com.azure.storage.queue.QueueAsyncClient;
@@ -25,22 +26,20 @@ import java.util.HashMap;
 import java.util.Map;
 
 /**
- * Azure Storage Queue template to support send / receive {@link Message} asynchronously.
+ * Azure Storage Queue template to support send / receive {@link Message}s asynchronously.
+ *
+ * You should checkpoint if message has been processed successfully, otherwise it will be visible again after certain
+ * time specified by {@link #receiveAsync(String, Duration)} }.
  */
-public class StorageQueueTemplate implements StorageQueueOperation {
+public class StorageQueueTemplate implements SendOperation {
     private static final Logger LOG = LoggerFactory.getLogger(StorageQueueTemplate.class);
-    private static final Duration DEFAULT_VISIBILITY_TIMEOUT_IN_SECONDS = Duration.ofSeconds(30);
     private static final String MSG_FAIL_CHECKPOINT = "Failed to checkpoint %s in storage queue '%s'";
-    private static final String MSG_SUCCESS_CHECKPOINT = "Checkpointed %s in storage queue '%s' in %s mode";
+    private static final String MSG_SUCCESS_CHECKPOINT = "Checkpointed %s in storage queue '%s'";
     private final StorageQueueClientFactory storageQueueClientFactory;
 
     private StorageQueueMessageConverter messageConverter = new StorageQueueMessageConverter();
 
-    private Duration visibilityTimeout = DEFAULT_VISIBILITY_TIMEOUT_IN_SECONDS;
-
     private Class<?> messagePayloadType = byte[].class;
-
-    private CheckpointMode checkpointMode = CheckpointMode.RECORD;
 
     /**
      * Create an instance using the supplied StorageQueueClientFactory.
@@ -48,48 +47,36 @@ public class StorageQueueTemplate implements StorageQueueOperation {
      */
     public StorageQueueTemplate(@NonNull StorageQueueClientFactory storageQueueClientFactory) {
         this.storageQueueClientFactory = storageQueueClientFactory;
-        LOG.info("StorageQueueTemplate started with properties {}", buildProperties());
+        LOG.info("StorageQueueTemplate started with default properties {}", buildProperties());
     }
 
     @Override
-    public <T> Mono<Void> sendAsync(String queueName, @NonNull Message<T> message,
-                                    PartitionSupplier partitionSupplier) {
+    public <T> Mono<Void> sendAsync(String queueName, @NonNull Message<T> message, PartitionSupplier partitionSupplier) {
         Assert.hasText(queueName, "queueName can't be null or empty");
         QueueMessageItem queueMessageItem = messageConverter.fromMessage(message, QueueMessageItem.class);
-        QueueAsyncClient queueClient = storageQueueClientFactory.getOrCreateQueueClient(queueName);
+        QueueAsyncClient queueClient = storageQueueClientFactory.createQueueClient(queueName);
         Assert.notNull(queueMessageItem, "queueMessageItem can't be null");
         return queueClient.sendMessage(queueMessageItem.getBody().toString()).then();
     }
 
-    @Override
-    public Mono<Message<?>> receiveAsync(String queueName) {
-        return this.receiveAsync(queueName, visibilityTimeout);
-    }
-
-    private Mono<Message<?>> receiveAsync(String queueName, Duration visibilityTimeout) {
+    /**
+     * Receive a message from the queue asynchronously.
+     * @param queueName the queue
+     * @param visibilityTimeout The timeout period for how long the message is invisible in the queue. If left empty
+     * the dequeued messages will be invisible for 30 seconds. The timeout must be between 1 second and 7 days.
+     * @return {@link Mono} of the next available {@link Message} or {@code null} if empty
+     */
+    public Mono<Message<?>> receiveAsync(String queueName, Duration visibilityTimeout) {
         Assert.hasText(queueName, "queueName can't be null or empty");
-
-
-        QueueAsyncClient queueClient = storageQueueClientFactory.getOrCreateQueueClient(queueName);
-
-
+        QueueAsyncClient queueClient = storageQueueClientFactory.createQueueClient(queueName);
         return queueClient.receiveMessages(1, visibilityTimeout)
             .next()
             .map(messageItem -> {
-
                 Map<String, Object> headers = new HashMap<>();
                 Checkpointer checkpointer = new AzureCheckpointer(() -> checkpoint(queueClient, messageItem));
-
-                if (checkpointMode == CheckpointMode.RECORD) {
-                    checkpointer.success().subscribe();
-                } else if (checkpointMode == CheckpointMode.MANUAL) {
-                    headers.put(AzureHeaders.CHECKPOINTER, checkpointer);
-                }
-
+                headers.put(AzureHeaders.CHECKPOINTER, checkpointer);
                 return messageConverter.toMessage(messageItem, new MessageHeaders(headers), messagePayloadType);
             });
-
-
     }
 
     private Mono<Void> checkpoint(QueueAsyncClient queueClient, QueueMessageItem messageItem) {
@@ -110,15 +97,8 @@ public class StorageQueueTemplate implements StorageQueueOperation {
     private Map<String, Object> buildProperties() {
         Map<String, Object> properties = new HashMap<>();
 
-        properties.put("visibilityTimeout", this.visibilityTimeout);
         properties.put("messagePayloadType", this.messagePayloadType);
-        properties.put("checkpointMode", this.checkpointMode);
-
         return properties;
-    }
-
-    private boolean isValidCheckpointMode(CheckpointMode checkpointMode) {
-        return checkpointMode == CheckpointMode.MANUAL || checkpointMode == CheckpointMode.RECORD;
     }
 
     private String buildCheckpointFailMessage(QueueMessageItem cloudQueueMessage, String queueName) {
@@ -126,8 +106,7 @@ public class StorageQueueTemplate implements StorageQueueOperation {
     }
 
     private String buildCheckpointSuccessMessage(QueueMessageItem cloudQueueMessage, String queueName) {
-        return String.format(MSG_SUCCESS_CHECKPOINT, StorageQueueHelper.toString(cloudQueueMessage), queueName,
-            checkpointMode);
+        return String.format(MSG_SUCCESS_CHECKPOINT, StorageQueueHelper.toString(cloudQueueMessage), queueName);
     }
 
     /**
@@ -147,21 +126,6 @@ public class StorageQueueTemplate implements StorageQueueOperation {
     }
 
     /**
-     * Get the {@code visibilityTimeoutInSeconds}.
-     * @return the visibilityTimeoutInSeconds.
-     */
-    public Duration getVisibilityTimeout() {
-        return visibilityTimeout;
-    }
-
-    @Override
-    public void setVisibilityTimeout(Duration visibilityTimeoutDuration) {
-        Assert.state(visibilityTimeoutDuration.isNegative() || visibilityTimeoutDuration.isZero(), "VisibilityTimeoutInSeconds should be positive");
-        this.visibilityTimeout = visibilityTimeoutDuration;
-        LOG.info("StorageQueueTemplate VisibilityTimeout becomes: {} seconds", this.visibilityTimeout.getSeconds());
-    }
-
-    /**
      * Get the {@code messagePayloadType}.
      * @return the messagePayloadType.
      */
@@ -169,25 +133,13 @@ public class StorageQueueTemplate implements StorageQueueOperation {
         return messagePayloadType;
     }
 
-    @Override
+    /**
+     * Set message payload type. Default is {@code byte[]}
+     * @param payloadType message payload type
+     */
     public void setMessagePayloadType(Class<?> payloadType) {
         this.messagePayloadType = payloadType;
         LOG.info("StorageQueueTemplate messagePayloadType becomes: {}", this.messagePayloadType);
     }
 
-    /**
-     * Get the {@code checkpointMode}.
-     * @return the {@code checkpointMode}.
-     */
-    public CheckpointMode getCheckpointMode() {
-        return checkpointMode;
-    }
-
-    @Override
-    public void setCheckpointMode(CheckpointMode checkpointMode) {
-        Assert.state(isValidCheckpointMode(checkpointMode),
-            "Only MANUAL or RECORD checkpoint mode is supported in StorageQueueTemplate");
-        this.checkpointMode = checkpointMode;
-        LOG.info("StorageQueueTemplate checkpoint mode becomes: {}", this.checkpointMode);
-    }
 }