+++ sdk/spring/spring-messaging-azure-storage-queue/CHANGELOG.md
@@ -3,9 +3,12 @@
 ## 4.0.0-beta.3 (Unreleased)
 
 ### Features Added
-
+- Support StorageQueueMessageConverter as a bean to support customize ObjectMapper.
 ### Breaking Changes
-
+- Remove `StorageQueueOperation`.
+- Remove configuration of checkpoint mode for StorageQueueTemplate, and support only MANUAL mode.
+- Remove auto creating Storage Queue when send/receive messages via `StorageQueueTemplate`.
+- Add the parameter of visibility timeout for StorageQueueTemplate#receiveAsync
 ### Bugs Fixed
 
 ### Other Changes